#refactor : divide script into components & script / combined creating form function (createFormElement)
