+++ script/task.js
@@ -1,4 +1,4 @@
-import CardForm, { onCancelEdit, onEdit } from "../components/cardForm.js";
+import createFormElement from "../script/formScript.js";
 import Modal from "../components/modal.js";
 import { renderTasks } from "./column.js";
 import { columns, handleDrag } from "./index.js";
@@ -18,7 +18,10 @@ export function createTask(task) {
 
     // 새 카드 컴포넌트 생성
     const newCard = document.createElement("li");
-    newCard.setAttribute("class", "card surface-default shadow-normal rounded-100");
+    newCard.setAttribute(
+        "class",
+        "card surface-default shadow-normal rounded-100"
+    );
     newCard.setAttribute("draggable", "true");
     newCard.innerHTML = taskHTML({ title, content });
     newCard.addEventListener("dragstart", (e) => handleDrag(e, task));
@@ -35,19 +38,9 @@ export function deleteTask(task) {
 }
 
 export function editTask(cardElement, task) {
-    const { title, content, created, column } = task;
-    cardElement.innerHTML = CardForm();
-    const inputs = cardElement.getElementsByTagName("input");
-    const [titleInput, contentInput] = inputs;
-    titleInput.value = title;
-    contentInput.value = content;
-    const form = cardElement.getElementsByTagName("form")[0];
-    form.addEventListener("submit", (e) => {
-        onEdit(e, task);
-    });
-    form.getElementsByTagName("button")[0].addEventListener("click", () =>
-        onCancelEdit(task, cardElement)
-    );
+    const formElement = createFormElement(task, undefined);
+    const parentNode = cardElement.parentNode;
+    parentNode.replaceChild(formElement, cardElement);
 }
 
 export function taskHTML({ title, content }) {
